+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 * Orders csv export has errors in the location column when orders are linked to holdings. Refs UIOR-921.
+* exportCsv is deprecated in stripes-utils. Refs UIOR-922.
 
 ## [3.1.0](https://github.com/folio-org/ui-orders/tree/v3.1.0) (2022-03-04)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v3.0.3...v3.1.0)